Рефакторинг кода перед мержем
