+++ Assets/Game/Spawners.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2d001b78fcd01bd42a3060a9d306d871
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
