+++ Assets/Game/Spawners/TreeSpawner.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4758d5cfb9325874ea174226b46988bb
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
