+++ Assets/Game/Spawners/TreeSpawnerConfig.asset
@@ -10,7 +10,7 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 38a65df0c6194c348bcc3fa01c4e1bc7, type: 3}
-  m_Name: TestConfig
+  m_Name: TreeSpawnerConfig
   m_EditorClassIdentifier: 
   Object: {fileID: 1749507651592192076, guid: 7c04d2cf1bf47e849bde1aa1c433a0e1, type: 3}
   Width: 40
