--- Assets/ISpawnable.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-/// <summary> Интерфейс, определяющий объект, который может быть уничтожен. </summary>
-public interface ISpawnable
-{
-    /// <summary> Событие, возникающее при уничтожении объекта. </summary>
-    public event Action<ISpawnable> OnObjectDestroyed;
-}
