+++ Assets/Scripts/ObjectSpawner/ISpawnable.cs.meta
