+++ Assets/Scripts/ObjectSpawner/ObjectSpawner.cs.meta
